update http services, move req/resp structures to http.model
